+++ flixel/src/org/flixel/system/FlxWindow.java
@@ -93,7 +93,7 @@ public class FlxWindow
 	 * @param BGColor		What color the window background should be, default is gray and transparent.
 	 * @param TopColor		What color the window header bar should be, default is black and transparent.
 	 */
-	public FlxWindow(String Title,float Width,float Height,boolean Resizable,Rectangle Bounds,long BGColor,long TopColor)
+	public FlxWindow(String Title,float Width,float Height,boolean Resizable,Rectangle Bounds,int BGColor,int TopColor)
 	{
 		super();
 		_width = (int) Width;
@@ -147,7 +147,7 @@ public class FlxWindow
 	 * @param Bounds		A rectangle indicating the valid screen area for the window.
 	 * @param BGColor		What color the window background should be, default is gray and transparent.
 	 */
-	public FlxWindow(String Title,float Width,float Height,boolean Resizable,Rectangle Bounds,long BGColor)
+	public FlxWindow(String Title,float Width,float Height,boolean Resizable,Rectangle Bounds,int BGColor)
 	{
 		this(Title, Width, Height, Resizable, Bounds, BGColor, 0x7f000000);
 	}
